New layout diagram, updated gui2 xml with scrollpane dynamix layout, added events to be comparable, added events time to be long instead of int, updated local storage to reflect long type change, broke up main, started sort

+++ src/events/Event.java
@@ -5,6 +5,6 @@ public interface Event {
 
 	public String getDescription();
 
-	public int getTime();
+	public long getTime();
 
 }
